+++ packages/view/src/components/FilteredAuthors/FilteredAuthors.scss
@@ -1,0 +1,9 @@
+.selected-container {
+  position: relative;
+  top: 30px;
+  display: flex;
+  flex-wrap: wrap;
+  width: 100%;
+  padding: 4px 6px;
+  box-sizing: border-box;
+}
