+++ packages/view/src/components/FilteredAuthors/index.ts
@@ -1,0 +1,1 @@
+export { default as FilteredAuthors } from "./FilteredAuthors";
