+++ packages/view/src/components/TemporalFilter/TemporalFilter.tsx
@@ -9,9 +9,6 @@ import BounceLoader from "react-spinners/BounceLoader";
 import { useGlobalData } from "hooks";
 import { throttle } from "utils";
 import type IDEPort from "ide/IDEPort";
-import { getInitData } from "components/VerticalClusterList/Summary/Summary.util";
-import { usePreLoadAuthorImg } from "components/VerticalClusterList/Summary/Summary.hook";
-import { Author } from "components/VerticalClusterList/Summary/Author";
 
 import {
   filterDataByDate,
@@ -34,7 +31,6 @@ const TemporalFilter = () => {
   const {
     data,
     filteredData,
-    selectedData,
     setFilteredData,
     filteredRange,
     setFilteredRange,
@@ -42,8 +38,6 @@ const TemporalFilter = () => {
     loading,
     setLoading,
   } = useGlobalData();
-  const authSrcMap = usePreLoadAuthorImg();
-  const selectedClusters = getInitData(selectedData);
 
   const loaderStyle: CSSProperties = {
     position: "fixed",
@@ -194,23 +188,6 @@ const TemporalFilter = () => {
       </div>
       <div className="line-charts" ref={wrapperRef}>
         <svg className="line-charts-svg" ref={ref} />
-      </div>
-
-      <div className="selected-cluster">
-        {authSrcMap &&
-          selectedClusters.map((selectedCluster) => {
-            return selectedCluster.summary.authorNames.map(
-              (authorArray: string[]) => {
-                return authorArray.map((authorName: string) => (
-                  <Author
-                    key={authorName}
-                    name={authorName}
-                    src={authSrcMap[authorName]}
-                  />
-                ));
-              }
-            );
-          })}
       </div>
     </article>
   );
